+++ src/DynamoRevit/ViewModel/RevitWatch3DViewModel.cs
@@ -236,7 +236,7 @@ namespace Dynamo.Applications.ViewModel
             try
             {
                 // we scale the tesselation rather than the curve
-                var conv = UnitConverter.DynamoToHostFactor(UnitType.UT_Length);
+                var conv = UnitConverter.DynamoToHostFactor(SpecTypeId.Length);
 
                 // use the ASM tesselation of the curve
                 var pkg = renderPackageFactory.CreateRenderPackage();