+++ src/Libraries/RevitNodes/Elements/Wall.cs
@@ -174,7 +174,7 @@ namespace Revit.Elements
                 throw new ArgumentNullException("wallType");
             }
 
-            height = height * UnitConverter.DynamoToHostFactor(UnitType.UT_Length);
+            height = height * UnitConverter.DynamoToHostFactor(SpecTypeId.Length);
 
             if (height < 1e-6 || height > 30000)
             {