+++ CHANGELOG.md
@@ -1,4 +1,22 @@
+## 0.2.7
+* Fix CI build error.
+
+## 0.2.6
+* Update some System tests and their dynamo files.
+* Rename "IncludeShadows" in node "Element.GetHostedElements".
+* Upgrade some obsolete classed and functions for RevitAPI updated
+
+## 0.2.5
+* Update RevitAPI folder name to net48.
+
+## 0.2.4
+* Update Assembly Shared Info for Revit Preview Release
+
+## 0.2.3
 * Element Type nodes - introduces an ElementType category to the Dynamo library
+* Added some new Category Nodes - Group nodes, Space nodes, Elevation marker nodes, Dimension nodes, Area nodes, View port nodes, Family document nodes, Family instance nodes, Warning nodes.
+* Element.Geometry should provide geometry for FamilySymbols with no instances
+* Sig file generation has been included in CICD of D4R
 
 ## 0.2.2
 * Update .Net to 4.8.